RocksDB (old) vs Pebble (new):

```
name                               old time/op    new time/op    delta
MVCCComputeStats/valueSize=8-16       183ms ± 1%     176ms ± 2%   -4.11%  (p=0.000 n=10+10)
MVCCComputeStats/valueSize=32-16      124ms ± 5%     107ms ± 2%  -13.95%  (p=0.000 n=10+9)
MVCCComputeStats/valueSize=256-16    58.4ms ± 3%    49.4ms ± 6%  -15.44%  (p=0.000 n=10+9)
MVCCFindSplitKey/valueSize=32-16     88.0ms ± 2%    81.8ms ± 1%   -7.06%  (p=0.000 n=10+10)

name                               old speed      new speed      delta
MVCCComputeStats/valueSize=8-16     366MB/s ± 1%   382MB/s ± 2%   +4.29%  (p=0.000 n=10+10)
MVCCComputeStats/valueSize=32-16    539MB/s ± 4%   627MB/s ± 2%  +16.18%  (p=0.000 n=10+9)
MVCCComputeStats/valueSize=256-16  1.15GB/s ± 2%  1.36GB/s ± 5%  +18.31%  (p=0.000 n=10+9)
MVCCFindSplitKey/valueSize=32-16    763MB/s ± 2%   821MB/s ± 1%   +7.60%  (p=0.000 n=10+10)
```

Release note: None